Create static website with Hugo
Adds the skeleton from the Hugo static site generator (gohugo.io),
the Hermit theme as a git submodule (https://themes.gohugo.io/hermit/),
and a couple of empty content pages.

The docs recommend setting up the theme as a submodule, even though I
abhore them, because we'll be deploying to Netlify.